# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SANOVIM — Marketing Médico com IA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /publishing/publish:
    post:
      operationId: publishPost
      tags: [publishing]
      summary: Publish a post immediately to Instagram
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublishPostRequest"
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
  /publishing/schedule:
    post:
      operationId: schedulePost
      tags: [publishing]
      summary: Schedule a post for later publishing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SchedulePostRequest"
      responses:
        "200":
          description: Scheduled post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledPost"
  /publishing/history:
    get:
      operationId: getPublishingHistory
      tags: [publishing]
      summary: Publishing history with status
      parameters:
        - name: account
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccountName"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Publishing history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublishedPost"
  /publishing/scheduled:
    get:
      operationId: getScheduledPosts
      tags: [publishing]
      summary: Get all scheduled (pending) posts
      parameters:
        - name: account
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccountName"
      responses:
        "200":
          description: Scheduled posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledPost"
  /publishing/scheduled/{id}:
    delete:
      operationId: cancelScheduledPost
      tags: [publishing]
      summary: Cancel a scheduled post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Video ─────────────────────────────────────────────────────────────
...
  /market/trends:
    get:
      operationId: getMarketTrends
      tags: [market]
      summary: Market trends for Brazilian digital medical market
      parameters:
        - name: niche
          in: query
          required: false
          schema:
            type: string
            enum: [tricologia, ortopedia, medicina_esportiva, baby_tracking, geral]
            default: geral
      responses:
        "200":
          description: Market trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketTrendsResponse"
...
components:
  schemas:
...
    AccountName:
      type: string
      enum: [drdaniel, angelica, loysby]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    PublishPostRequest:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        caption:
          type: string
        hashtags:
          type: string
        image:
          type: string
          format: binary
      required: [account, caption]
    SchedulePostRequest:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        caption:
          type: string
        hashtags:
          type: string
        scheduledAt:
          type: string
          format: date-time
        image:
          type: string
          format: binary
      required: [account, caption, scheduledAt]
    PublishResult:
      type: object
      properties:
        success:
          type: boolean
        instagramPostId:
          type: string
        permalink:
          type: string
        publishedAt:
          type: string
          format: date-time
      required: [success, publishedAt]
    ScheduledPost:
      type: object
      properties:
        id:
          type: integer
        account:
          $ref: "#/components/schemas/AccountName"
        caption:
          type: string
        hashtags:
          type: string
        imageUrl:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, published, failed, cancelled]
        createdAt:
          type: string
          format: date-time
      required: [id, account, caption, scheduledAt, status, createdAt]
    PublishedPost:
      type: object
      properties:
        id:
          type: integer
        account:
          $ref: "#/components/schemas/AccountName"
        caption:
          type: string
        imageUrl:
          type: string
        instagramPostId:
          type: string
        status:
          type: string
          enum: [published, failed, scheduled, cancelled]
        publishedAt:
          type: string
          format: date-time
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, account, caption, status, createdAt]
...
    MarketTrendsResponse:
      type: object
      properties:
        niche:
          type: string
        trends:
          type: array
          items:
            $ref: "#/components/schemas/MarketTrendItem"
        analysis:
          type: string
        fetchedAt:
          type: string
          format: date-time
      required: [trends, fetchedAt]
    MarketTrendItem:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        opportunity:
          type: string
        growthPotential:
          type: string
          enum: [low, medium, high, very_high]
        sources:
          type: array
          items:
            type: string
      required: [title, description, growthPotential]
...